+++ examples/cuda/plotting.cu
@@ -18,10 +18,10 @@
 const unsigned DIMX = 1000;
 const unsigned DIMY = 800;
 
-static const float    dx = 0.1;
+static const float    dx = 0.1f;
 static const float    FRANGE_START = 0.f;
 static const float    FRANGE_END = 2 * 3.141592f;
-static const size_t   DATA_SIZE = ( FRANGE_END - FRANGE_START ) / dx;
+static const size_t   DATA_SIZE = (size_t)(( FRANGE_END - FRANGE_START ) / dx);
 
 void kernel(float* dev_out, int functionCode);
